Change logs documents to better work with CosmosDB
The default is now to store timestamps using UTC. UTC timestamps are now formatted exactly as CosmosDB does it to allow range queries on timestamps. The current culture does not affect the format of UTC timestamps. The log event level is now numeric to make it easy to query for events above a specific level.